Removed deprecated class JUnit44Runner
