+++ src/test/java/org/mockitousage/junitrunner/JUnitRunnerTest.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2007 Mockito contributors
+ * This program is made available under the terms of the MIT License.
+ */
+package org.mockitousage.junitrunner;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertSame;
+import static org.mockito.Mockito.verify;
+import static org.mockitousage.junitrunner.Filters.methodNameContains;
+
+import java.util.List;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+@SuppressWarnings( { "unchecked" })
+public class JUnitRunnerTest {
+
+    @InjectMocks
+    private ListDependent listDependent = new ListDependent();
+
+    @Mock
+    private List list;
+
+    @Test
+    public void shouldInitMocksUsingRunner() {
+        list.add("test");
+        verify(list).add("test");
+    }
+    @Test
+    public void shouldInjectMocksUsingRunner() {
+        assertSame(list, listDependent.getList());
+    }
+
+    @Test
+    public void shouldFilterTestMethodsCorrectly() throws Exception{
+    	MockitoJUnitRunner runner = new MockitoJUnitRunner(this.getClass());
+
+        runner.filter(methodNameContains("shouldInitMocksUsingRunner"));
+
+        assertEquals(1, runner.testCount());
+    }
+
+    class ListDependent {
+        private List list;
+
+        public List getList() {
+            return list;
+        }
+    }
+}
